Remove the "-formatted" field
Removes the pre-formatted version of the field. This functionality was
provided by HOF's old date controller but we now believe that it is not
the job of the controller to provide this as the functionality
ultimately belongs in the view. In fact, the functionality for this
already exists in hmpo-template-mixins.